Fixing all Pylint recommandations
