+++ sdk/communication/azure-communication-callautomation/azure/communication/callautomation/_events.py
@@ -21,7 +21,10 @@ from ._generated.models import (
     CallParticipant
 )
 
-from ._models import deserialize_identifier
+
+from ._shared.models import (
+    deserialize_identifier
+)
 
 
 class AddParticipantSucceeded(AddParticipantSucceededRest):
