+++ sdk/communication/azure-communication-callautomation/pyproject.toml
@@ -1,0 +1,5 @@
+[tool.azure-sdk-build]
+mypy = false
+pyright = false
+type_check_samples = false
+verifytypes = false
