+++ sdk/communication/azure-communication-callautomation/sdk_packaging.toml
@@ -1,7 +1,7 @@
 [packaging]
 auto_update = false
 package_name = "azure-communication-callautomation"
-package_pprint_name = "Communication CallAutomation"
+package_pprint_name = "Communication Call Automation"
 package_doc_id = ""
 is_stable = false
 is_arm = false
